+++ circuit_definitions/src/circuit_definitions/aux_layer/compression_modes/mod.rs
@@ -62,17 +62,17 @@ pub mod mode_3;
 pub mod mode_3_for_wrapper;
 pub mod mode_4;
 pub mod mode_4_for_wrapper;
-pub mod mode_to_l1;
-pub mod mode_to_l1_for_wrapper;
+pub mod mode_5;
+pub mod mode_5_for_wrapper;
 
 pub use self::mode_1::*;
 pub use self::mode_2::*;
 pub use self::mode_3::*;
 pub use self::mode_4::*;
-pub use self::mode_to_l1::*;
+pub use self::mode_5::*;
 
 pub use self::mode_1_for_wrapper::*;
 pub use self::mode_2_for_wrapper::*;
 pub use self::mode_3_for_wrapper::*;
 pub use self::mode_4_for_wrapper::*;
-pub use self::mode_to_l1_for_wrapper::*;
+pub use self::mode_5_for_wrapper::*;
