+++ src/data_source/local_file_data_source.rs
@@ -17,9 +17,9 @@ use snark_wrapper::franklin_crypto::bellman::plonk::better_better_cs::proof::Pro
 use snark_wrapper::franklin_crypto::bellman::plonk::better_better_cs::setup::Setup as SnarkSetup;
 use snark_wrapper::franklin_crypto::bellman::plonk::better_better_cs::setup::VerificationKey as SnarkVK;
 
-use std::{error::Error, fs::File};
-use std::sync::Arc;
 use derivative::*;
+use std::sync::Arc;
+use std::{error::Error, fs::File};
 
 #[derive(Derivative)]
 #[derivative(Clone, Copy, Debug)]
@@ -184,11 +184,10 @@ impl SetupDataSource for LocalFileDataSource {
         ))
         .map_err(|el| Box::new(el) as Box<dyn Error>)?;
 
-        let result = Arc::new(
-            SnarkSetup::read(&mut file).map_err(|el| Box::new(el) as Box<dyn Error>)?
-        );
+        let result =
+            Arc::new(SnarkSetup::read(&mut file).map_err(|el| Box::new(el) as Box<dyn Error>)?);
 
-        let result = ZkSyncSnarkWrapperSetup::from_inner(circuit_type,result);
+        let result = ZkSyncSnarkWrapperSetup::from_inner(circuit_type, result);
 
         println!("Wrapper setup read from file. Took {:?}", start.elapsed());
